# src/main/resources/public/docs/openapi.yml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: Pulmonis API
...
paths:
...
  /public/schooling/{schoolingId}:
    get:
      tags:
        - Public Schooling
      summary: Get schooling by id
      parameters:
        - in: path
          name: schoolingId
          required: true
          schema:
            type: number
            format: int64
            minimum: 1
          description: The schooling ID
      description: Get schooling for visitors which is published. No authentication needed.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolingResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
...
  /user/listing:
    get:
      tags:
        - Users management
      summary: Get list of all users
      description: Only user with admin priveleges can view all users.
      parameters:
        - in: path
          name: page
          schema:
            type: integer
            default: 0
            description: Parameter for specifing page of users to be retrieved. If not provided, page=0 will be returned. Default size of page is 16 and elements sorted by update time.
            example: 3
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListingResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /user/update-status/{email}:
    put:
      tags:
        - Users management
      summary: Update user's status
      description: Only user with admin priveleges can update users status
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
          description: Person's email which is email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: Status of user.
                  enum: [ active, inactive, pending ]
                  example: active
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /schooling/create:
    post:
      tags:
        - Schooling
      summary: Create a new schooling.
      description: Create a new schooling. User ID is taken from JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolingRequest"
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolingResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        412:
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /schooling/listing:
    get:
      tags:
        - Schooling
      summary: List schoolings by a user's ID
      description: List all user's schoolings. User ID is retrieved from JWT token.
      parameters:
        - in: path
          name: page
          schema:
            type: integer
            default: 0
            description: Parameter for specifing page of schoolings to be retrieved. If not provided, page=0 will be returned. Default size of page is 16 and elements sorted by update time.
            example: 3
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolingListingResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /schooling/{schoolingId}:
    get:
      tags:
        - Schooling
      summary: Specific schooling by schooling's id
      description: Return doctor's schooling by schooling's id. User ID is retrieved from JWT token.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolingResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /schooling/remove/{schoolingId}:
    delete:
      tags:
        - Schooling
      summary: Removing schooling by its ID
      description: User can remove schooling by it. User ID is retrieved from JWT token.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoveSchooling"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /schooling/edit/{schoolingId}:
    patch:
      tags:
        - Schooling
      summary: Edit schooling.
      description: Edit schooling. Schooling id is retrieved from path. User id is retrieved from JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolingRequest"
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolingResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        412:
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    SchoolingRequest:
      type: object
      properties:
        title:
          type: string
          description: Title of the event
          example: Title
        description:
          type: string
          description: Description of the event
          example: Description
        address:
          type: string
          description: Address where schooling takes place.
          example: Lorem
        city:
          type: string
          description: City where schooling takes place.
          example: Ipsum
        category:
          type: string
          description: Category of schooling.
          example: Lorem ipsum
        eventDateTime:
          type: date-time
          description: Date time of the event in the format yyyy-mm-ddThh:mm:ss
          example: 2007-12-03T10:15:30
        registrationDeadline:
          type: date-time
          description: Schooling event deadline in format of yyyy-mm-ddThh:mm:ss
          example: 2015-12-03T10:15:35
        free:
          type: boolean
          description: Is event free or not
          example: false
        price:
          type: double
          description: Participation price for the event
          example: 20.0
        published:
          type: boolean
          description: Whether the schooling is published or not.
          example: false
...
    SignupResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            email:
              type: string
              description: Person's email which is email
              example: testing@ut.ee
            status:
              type: string
              description: User's status
              example: pending
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
...
    UserListingResponse:
      type: object
      properties:
        data:
          properties:
            totalItems:
              type: integer
              example: 1
              description: Total items returned.
            users:
              type: array
              items:
                $ref: "#/components/schemas/UserData"
            totalPages:
              type: integer
              example: 1
              description: Total pages.
            currentPage:
              type: integer
              example: 0
              description: Currently selected page.
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
    UserData:
      type: object
      properties:
        email:
          type: string
          description: Person's email which is email
          example: pulmonis@eesti.ee
        status:
          type: string
          description: User's status.
          example: pending
    SchoolingResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/SchoolingData"
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
    SchoolingListingResponse:
      type: object
      properties:
        data:
          properties:
            totalItems:
              type: integer
              example: 1
              description: Total items returned.
            schoolings:
              type: array
              items:
                $ref: "#/components/schemas/SchoolingData"
            totalPages:
              type: integer
              example: 1
              description: Total pages.
            currentPage:
              type: integer
              example: 0
              description: Currently selected page.
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
    RemoveSchooling:
      type: object
      properties:
        data:
          properties:
            schoolingId:
              type: integer
              format: int64
              description: Schooling ID.
              example: 7
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
    SchoolingData:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Schooling ID.
          example: 7
        title:
          type: string
          description: Title of schooling.
          example: Ipsum Lorem
        description:
          type: string
          description: Description of schooling.
          example: Lorem ipsum
        address:
          type: string
          description: Address where schooling takes place.
          example: Lorem
        city:
          type: string
          description: City where schooling takes place.
          example: Ipsum
        category:
          type: string
          description: Category of schooling.
          example: Lorem ipsum
        eventDateTime:
          type: date-time
          description: Schooling event date time in format of yyyy-mm-ddThh:mm:ss
          example: 2015-12-03T10:15:35
        free:
          type: boolean
          description: Whether the schooling is free of charge or not.
          example: false
        price:
          type: double
          description: The price of schooling if schooling not free of charge. Otherwise price defaults to 0.0.
          example: 420.69
        registrationDeadline:
          type: date-time
          description: Schooling event deadline in format of yyyy-mm-ddThh:mm:ss
          example: 2015-12-03T10:15:35
        published:
          type: boolean
          description: Whether the schooling is published or not.
          example: true
...
    EmptyError:
      type: array
      nullable: true
      description: List of errors returned by API.
      example: null
    SuccessfulStatus:
      type: integer
      example: 200
      default: 200
      description: Successful response from API.
    UnsuccessfulStatus:
      type: integer
      example: 400
      description: Successful status response from API. Can vary from 400 to 500.
    UnsuccessfulResponse:
      type: object
      properties:
        data:
          type: object
          nullable: true
          example: null
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
        status:
          $ref: "#/components/schemas/UnsuccessfulStatus"
    ApiError:
      type: object
      properties:
        message:
          type: string
          description: Error message returned from API.
          nullable: true
          example: "string"
        code:
          type: string
          description: Error code returned from API.
          example: "string"
security:
  - bearerAuth: [ ]